login: Stop ARCVM and related Upstart job on Chrome exit
DD: go/arcvm-shutdown-on-chrome-exit

When ash-Chrome restart within a user session happens (e.g., on crash or
user-initiated restart), we need to shut down the stale ARCVM instance.
We currently do it in the restarted Chrome session, but this has caused
bugs due to the unexpected interaction between stale ARCVM and the
restarted services that interacts with ARCVM. To avoid it in the future,
this CL introduces logic to shut down ARCVM and some ARCVM-related
Upstart services that needs to be restarted along with it in
SessionManagerService::HandleExit(), only when it is about to relaunch
new ash-Chrome.

BUG=b:290194650
TEST=FEATURES=test emerge-<BOARD> chromeos-login
TEST=Manually trigget ash-Chrome restart with chrome://restart or
TEST=chrome://inducebrowsercrashforrealz and check in /var/log/messages
TEST=that ARCVM and the related jobs are stopped before the new
TEST=ash-Chrome is launched.

Change-Id: I42029cf62f598138433fe54fe50b84a3bc956c10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4863022
Commit-Queue: Momoko Hattori <momohatt@chromium.org>
Tested-by: Momoko Hattori <momohatt@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>