shill: portal: separate probe rotation and delay calculation
This patch separates in PortalDetector the state associated with probe
url rotation from the state associated with delay calculation when
scheduling a retry.

It is now possible to reset the exponential backoff without impacting
the probe rotation logic. This is useful for scheduling a quick retry of
portal detection without also resetting the probe URLs used.

BUG=b:300089731
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: Ib811d1af91a5dcf95df778a5adea312e1e827a4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4886033
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>