installer: Fix numbering of enum UMA
This UMA has been broken for a year, and has never actually sent data.
It's been tripping over a validation step in libmetrics checking that
the max value for an enum metric is > 1.

It's safe to change this because the data has never actually been sent
to the server, so there's no mismatch.

Libmetrics doesn't currently provide a testing implementation that does
the same validation as the real implementation, so there's no good way
to write tests to catch this sort of thing.

BUG=b:297236244
TEST=cros_workon_make -breven --test chromeos-installer

Change-Id: I65e33171bc8c8121e1c089bfd695b2155cb333f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4877719
Auto-Submit: Ted Brandston <tbrandston@google.com>
Tested-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Ted Brandston <tbrandston@google.com>
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>