cryptohome: Replace UserSecretStash with DecryptedUss
This change rewrites all the uses of UserSecretStash to instead use
DecryptedUss, including using Transaction for all modifications. As a
consequence it can also get rid of UserSecretStash itself, although it
doesn't get rid of the entire header since the flag and experiment
manipulation functions are still there and they don't really make sense
to move.

For the most part this change is just a mechanical rewrite, replacing
factory function calls with their DecryptedUss equivalent, replacing
unique_ptr<UserSecretStash> with optional<DecryptedUss> (or just
DecryptedUss), and rewriting transactional code to use transactions
instead of snapshot+cleanup. However, in a few cases where existing
modifications were being done in an non-transactional way, it does have
to fix them up.

In particular this fixes UpdateAuthFactor to use transactions, and a
minor consequence of this is that all the in-memory USS operations are
moved to after the auth factor update portion of the change.

There are also a few other cases where some functions were doing
modifications without any transactional behavior. Depending on the
context I either changed them to do a transaction, or to take a
Transaction& as a parameter.

This completed the largest part of the rewrite. There are still some
followup cleanups that need to happen, to simplify and get rid of any
leftover transitional code, but this is what USS code should mostly look
like going forward.

BUG=b:295898583
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ib18d4e79fd733eb50518ed117c6d2162dcc9145e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4887214
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>