patchpanel: Always turn multicast flag on for ARC container
http://crrev.com/c/4834903 changes the multicast flag for ARC
container's interfaces from:
- following the physical interfaces multicast flag,
to:
- always default to off.

This is a regression as we expect multicast for Android to always be
supported. This expectation is enforced by CTS (see b/299992183) and
CDD (starting on U).

This patch changes the behavior for ARC container multicast flag to
always on. This is preferred because:
- Android expect the flag.
- Reduces behavior divergence with ARCVM as it always has the flag on.
- Allows more straightforward way of disabling multicast on the host
  without affecting ARC.

BUG=b:299992183
BUG=b:301045084
TEST=unit test

Change-Id: I51ef1ea862f7e31cc9eecf346bd734da79bd785c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4881518
Tested-by: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jason Iman <jasongustaman@chromium.org>