shill: Add EnableDHCPQoS property on Manager
Add a new property on Manager to control the QoS behavior for dhcpcd.
Note that on a non-dev machine, this property will only be set by Chrome
when the user session changed based on the flag values, so this property
won't be persisted into profile to make sure its value is synced with
the flag.

BUG=b:296959012
TEST=FEATURES="test" emerge-$BOARD shill
TEST=Ran `dbus-send --system --dest=org.chromium.flimflam --print-reply
  / org.chromium.flimflam.Manager.SetProperty string:"EnableDHCPQoS"
  boolean:true` on DUT, observed that the new egress DHCP packets have
  tos=0xc0.

Change-Id: I965b885cac6a2d61304a230fb49375a3e6ee14af
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4881786
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Damien Dejean <damiendejean@google.com>