installer: Update EFI Boot Management tests
In efi_boot_management_test.cc we're using NiceMock and checking that
things aren't called. That's what StrictMock is supposed to cover,
though.

Use a StrictMock and merge some of the metric-specific testing into the
rest of the tests.

Proposed in https://crrev/c/4877720

BUG=None
TEST=cros_workon_make -breven --test chromeos-installer

Change-Id: I6f81ca34edcd45633e552a11649aecc8988ec6f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4881441
Tested-by: Ted Brandston <tbrandston@google.com>
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Ted Brandston <tbrandston@google.com>