hiberman: Keep track of user logouts
The metrics indicate that one of the most common reasons for a
hibernate (suspend) failure is that the hiberimage does not
exist. This is only expected to happen if the user logged out
previously, which results in tearing down the hibernate. The
image is not recreated on successive logins.

Report a dedicated event when the hiberimage does not exist
because the user logged out, to be able to distinguish this
condition from possible others.

BUG=b:301443990
TEST=log in
  hiberman teardown-hiberimage
  hiberman hibernate
    => fails
  chrome://histograms/hibernate.abort
    => value 5 (UserLoggedOut) was reported

Change-Id: I982ff34a2dcc7acf501a027be33add900b9bcf20
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4884431
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>