lorgnette: Expose SetOption from SaneDevice
SaneDeviceImpl already contains a SetOption function that it uses to set
all of the known options from the Scan app.  The new d-bus API allows
setting arbitrary options, so hoist this up to SaneDevice and connect it
to the new protobuf SaneOption::Set function.

BUG=b:274860786
TEST=CQ

Change-Id: I96182e908b854aa79358ecbfbf31174ff3088619
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4899031
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>