vm_tools: Refactors sommelier-output to batch up and group send updates
This is a pre-patch for the introduction of xdg-output. By batching up
the output/update information inside one function it'll make it easier
to integrate the logic of forwarding output and direct scale together.

Also added some comments to help explain what all of the various scale
numbers are in sl_context and sl_host_output.

BUG=b:289158040
TEST=Unittests run, tested on device.

Change-Id: Ib43853af6ef5a11ea0d4a6dcd28293333f443140
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4891392
Reviewed-by: Chloe Pelling <cpelling@google.com>
Commit-Queue: Justin Huang <justinhuang@google.com>
Tested-by: Justin Huang <justinhuang@google.com>