shill: cellular: allow aborting an ongoing tethering network acquisition attempt
The tethering disconnection logic implemented for both the "DUN as
DEFAULT" and "Multiplexed DUN" cases is already capable of recovering
the state regardless of when they are called, and therefore these
methods can also be used to abort an ongoing connection attempt.

Whenever a request to stop the tethering session happens, the
tethering manager will call ReleaseTetheringNetwork() unconditionally,
even if there is no tracked Network yet. If such a request without a
Network arrives the cellular object, it will be treated as an early
abort request.

Upon an abort, we will first complete the original tethering network
acquisition  request with an "OperationAborted" error, and then launch
a clean tethering network release, which would recover the original
state before the tethering attempt started. The unit tests are updated
to take into account this change; before we were waiting for the
recovery to finish before completing the acquisition attempt with an
error.

The tethering network release operations are not abortable. If the
tethering manager timeout expires before a release has finished, the
operation will still go on. The logic may not allow to request a new
tethering attempt until that release logic has really finished in the
cellular side.

BUG=b:295331134
TEST=Enable/disable hotspot manually

Change-Id: I551d1627191df71aa4cd788fafafc02811579fa8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4889621
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Tested-by: Aleksander Morgado <aleksandermj@google.com>