hiberman: Log the errno as metric when image creation or saving fails
Log the errno reported by the kernel as a metric when the hibernate
image can not be created or stored to disk.

The purpose of the errno metric is to learn which are the prevalent
errors when hibernating, these might not always be seen with local
testing. Once these errors are identified they can be added as enums
to the Platform.Hibernate.Abort metric (if needed) and the errno
metric can be deprecated.

BUG=b:301462609
TEST=force hibernate error and check chrome://histograms

Change-Id: I342f906ce2ebcf3614b8263bb1e7e2fc236f53dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4892626
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>