cryptohome: Fix excessive policy file logs
Currently the use of configured policy files is rare and as a
result, most of the code paths will reach a non-existent policy
file, this will result in an excessive log of not finding policy
files. This CL removes the one log that happens very frequently in
SendAuthFactorStatusUpdateSignal in order to limit this.

BUG=b:302010393
TEST=none

Change-Id: I807d06d78342e5512803e8dd6632a58c895181ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4894604
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Behnood Momenzadeh <behnoodm@chromium.org>
Commit-Queue: Behnood Momenzadeh <behnoodm@chromium.org>