hiberman: Skip resume if user ids don't match
Don't try to resume from hibernate if the account id of the user
who authenticated is not the same as the one of the user who
hibernated. To determine that we store the obfuscated account id
of the hibernating user on the stateful partition. At resume time
we get the account id from Chrome and compare it with that on
disk. If the two don't match resume from hibernate is skipped.

BUG=b:301600911
TEST=hibernate resume is skipped if another user logs in
  hibernate resume still works if the same user logs in

Cq-Depend: chromium:4908322
Change-Id: Ifd806fc455d965a488d64714e537fad0a73d4b9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4899024
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>