hammerd: hammertests: update wait time for improving test stability
The hammer tests uses delay to wait for the base keyboard to become
ready, or wait for the base keyboard is ready in RW firmware. The wait
time was set as 2 seconds, which is at the edge for some AP and USB
peripherals on AP side to successfully recognized a new USB device is
reconnected.

Add or Update the wait time by using a much safer waiting time (2x of
the original waiting time), and extract the jump to RW ready time as a
variable for clarifying the intention of waiting, and for any future
changes.

BUG=b:278138283
TEST=run complete hammer tests with `./hammertests_control.py` on DUT
     and passed.

Change-Id: I647e72e5e6dd80d803eecba53a75b6399cce696e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4895106
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Tested-by: Sung-Chi Li <lschyi@chromium.org>
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>