patchpanel: Wrap qos_detect chain in another static chain
Previously, we install the jump rules to the qos_detect chain in the
mangle table dynamically. Since the jump rules are appended to the
end of the OUTPUT/PREROUTING chain, sometimes it might be shortcut by
the ACCEPT rule in the skip_apply_vpn_mark chain.

This patch modifies the setup to wrap the qos_detect chain in another
static chain (qos_detect_static). The jump rules to this chain are
installed statically in Datapath::Start() and thus the rule order in the
mangle table will be fixed. qos_detect_static will hold the dynamic jump
rule to qos_detect which will be installed/removed when
enabling/disabling QoS feature.

BUG=b:301566901
TEST=FEATURES="test" emerge-$BOARD patchpanel
TEST=tast run $DUT network.QosNetworkControl

Change-Id: Icd740a03ba55bd0a83a809f5e31d35707a5346e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4896085
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Damien Dejean <damiendejean@google.com>