diagnostics: Improve UnexpectedRoutineExceptionCallback
The old implementation of UnexpectedRoutineExceptionCallback used to
raise fatal failures of GoogleTest. However, it's common to use RunLoop
for testing routine v2 but the RunLoop won't stop when a fatal failure
was raised. Therefore, when an unexpected routine exception was raised,
the unit tests just hang without providing any message, which was
confusing for developers.

This commit changes the fatal GoogleTest failure to the CHECK failure,
which will stop the RunLoop and provide a clear message to developers.

BUG=none
TEST=FEATURES=test emerge-volteer diagnostics

Change-Id: I261221973a4ae7844fe6df3030d3b39d2a59a25a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4896090
Tested-by: Wei-Luan Wang <weiluanwang@google.com>
Commit-Queue: Wei-Luan Wang <weiluanwang@google.com>
Reviewed-by: Byron Lee <byronlee@chromium.org>