shill: portal: complete trial sooner if redirected
Some captive portals will silently block HTTPS traffic until the user
signs in, causing the HTTPS probe used in network validation to timeout.
It is possible to avoid waiting for the complete timeout duration and to
complete network validation sooner if the HTTP probe gets redirected
with a Location header.

This patch also improves logging when receiving a result for a HTTP
probe and clarifies which cases are valid redirect results and which
cases are incomplete results.

Manual tests: setup http server that sends a static 302 result to all
request on DUT2, pointed DUT1's HTTP probes to DUT2 and blocked HTTPS
traffic with iptables. Observed network validation completes on the HTTP
redirect.

BUG=b:302427165
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: Icdf3529bfd0c04715c3fdd0aea9018d0875bdda2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4900118
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>