lorgnette: Parse ScannerOption into SaneOption
SaneOption::ToScannerOption turns a SaneOption into a ScannerOption
protocol buffer.  This CL adds a Set() overload that does the opposite.
This will allow incoming d-bus SetOptions calls to be turned into SANE
backend updates.

Because the new d-bus API allows clients to request automatic setting
for options that support it, also track whether an option is supposed to
be set automatically or to a specific value.

BUG=b:274860786
TEST=unit tests

Change-Id: If82be363dbf06d3f378ff4da2217e1309eaae08d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4899030
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>