modemfwd: em060 firmware handling in upload script
Add new modem to fw upload script. The expectation for this vendor is
that firmware drops will have a directory structure that matches the
structure we ship onto the device, i.e.

{MAIN,CARRIER,OEM}/{version}/{main,carrier,oem}.bin

and the supporting engineer will run this script on each file in the
structure (that is not already in GS). This should be intuitive for the
vendor, and will allow the script to extract the appropriate versioning
for each file from the directory structure.

Change was validated by running some sample main and carrier files
through the script, validating that the nested directory structure is
correct before being compressed.

BUG=b:291927162
TEST=Ran above tests

Change-Id: I614ca90695e4bd7e313b4027f8bd4b035d489b13
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4898108
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Madhav <madhavadas@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>