cryptohome: Get rid of the clobber option for USS
This gets rid of the clobber parameter for AddAuthFactorToUssTransaction
and the associated enum (of which this was the last remaining use).

The parameter isn't really very useful anymore as the only call site
that uses clobber=disabled is in the Replace implementation where it's
just double-checking something that we already checked in the Replace
precondition. It's not really worth supporting the complexity of the
branching just for that extra double-check.

BUG=b:295898583
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ia8f7ea2746bb52680b113f7881e4bef2fff022f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4903282
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>