cryptohome: Log mount type in the namespace mounter
The mounter process sometimes spends a long time in `PerformMount`,
which can cause the parent process to time out and send a crash
report. Add a log of the mount type so that in future crash reports we
can see if the problem correlates with the mount type.

BUG=b:300839936
TEST=emerge-reven cryptohome
TEST=deploy to dut, reboot, login, verify new log shows up

Change-Id: I912e3744589b2c946e3d391cd37b81d280fc3b25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4902467
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Betul Soysal <betuls@google.com>