libbrillo: treat deferred remove as non-remove in fake device mapper.
This is necessary for tests on cryptohome's dmcrypt based vault
to pass. Without this change, the vault sets deferred remove
of dmcrypt devices, which leads the immediate removal of the
devices. That causes suspend and key evict operation
on devices to fail since the specified devices have gone.

BUG=b:294127303
TEST=CQ

Change-Id: I9099aa198a5f1576aa15022aeb26deb2c8975a27
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4905265
Commit-Queue: Zi Lin <lziest@google.com>
Tested-by: Zi Lin <lziest@google.com>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>