arc: Disable flag based stop of keymaster/keymint
Earlier, we were starting both the daemons - Keymaster
and KeyMint, and needed to stop 1 of them based on a flag.
Now, we are starting only of the 1 daemons on-demand
by D-Bus. Hence, flag based stopping of daemon is no
longer required

This CL removes the mechanism of stopping both the
daemons based upon the flag.

BUG=b:302645472
TEST=On-demand starting of Keymaster/KeyMint still works.

Change-Id: Ic4d0b6dc13cabf66a714abd8d6ae4e3b2f4f62a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4904848
Commit-Queue: Vaibhav Raheja <vraheja@chromium.org>
Reviewed-by: Yao Li <yaohuali@google.com>
Tested-by: Vaibhav Raheja <vraheja@chromium.org>