crash: Support fwdump generation in prod build in crash-reporter
This patch adds support for fwdump in crash-reporter. This patch
fetches user policy and logged in username. If the user is in an
allowlist or a googler and also policy flag is set to capture fw
dump then fw dump is collected in user's cryptohome directory.
The details of the design changes in crash-reporter is provided
in design doc:go/wifi-fwdumps-in-crash-reporter

BUG=b:291344512
TEST=FEATURES=test emerge-brya crash-reporter
TEST=Verified end to end through tast test
TEST=Validated fwdump generation in modified version of CL
that stores fwdump in crash-reporter's cryptohome directory
using tast test(WIP) crrev/c/4939167.
TEST=tast run <dut_IP> wifi.PolicyFwdumpCheck

Cq-Depend: chromium:5005503
Change-Id: I92f8ba79bc7ed704b89ee4dda308a6e825e05704
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4907501
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: Abhishek Kumar <kuabhs@chromium.org>
Commit-Queue: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>