camera: Update 99.camera.rules for access to mtk_hcp driver
In order to close source part of driver operation, MTK adds mtk_hcp
driver, which allows the driver calls back to userspace for certain
operations. Current approach is that MTK should migrate the part of
userspace codes into SCP, instead of calling back to userspace.

This patch is a temporary exception for the camera service to access
the driver before MTK finish the migration. It should be removed when
the SCP is working.

BUG=b:277309661
TEST=test on DUT

Change-Id: I4fe2fdb51756e7b601d6f5dad40dda97b40559c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4902874
Reviewed-by: Harvey Yang <chenghaoyang@chromium.org>
Tested-by: Hanlin Chen <hanlinchen@google.com>
Reviewed-by: becker hsieh <beckerh@chromium.org>
Commit-Queue: Hanlin Chen <hanlinchen@google.com>