lorgnette: Handle cancelling completed scans
In crrev.com/c/4905263, lorgnette stopped tracking completed jobs to
avoid triggering unnecessary cancel events.  However, this caused
lorgnette to start reporting CancelScan on a completed job as
already-cancelled.  The difference doesn't really matter from the device
point of view (the job is ended either way), but the
OPERATION_RESULT_CANCELLED response is confusing for the client.
Instead, check for the specific case of cancelling a completed job and
report success for it.

BUG=b:297442216
TEST=tast run scanner.AdvancedScan.*

Change-Id: I5493e447562948fd250f8c57272f5df03b48b601
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4907589
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Nathan Muggli <nmuggli@google.com>