crash: Add a "spool full" sentinel file.
It's been a persistent point of confusion with developers when we drop
crashes due to a full spool directory. We do log a message in
/var/log/messages, but that doesn't help if you don't know to look for
it (or if there's an ill-timed log rotation).

So, touch a file in the affected spool directory every time we drop
crashes for that reason to make it slightly more obvious.

BUG=b:186651955
TEST=tast run forward_crbook crash.User.max_enqueued_crash;\
     manually fill directory when not in test, verify file is left

Change-Id: I086450bbeca41a63d749feed608f927c0fbab5d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4902621
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>