shill: fix the rtnl_handler_fuzzer
This CL fixes the use-of-unitialized-value error at rtnl_handler_fuzzer.
The issue is caused by accessing the byte buffer outside the boundary.

This CL adds the size check before casting a raw buffer to a struct, and
uses the signed integer for RTA_NEXT() macro which performs subtraction
on the length of buffer.

BUG=b:302720665
TEST=Run rtnl_handler_fuzzer for 10 mins
TEST=FEATURES=test emerge-nissa shill-net

Change-Id: I1d444898e4cc45580a3e9d972362777c7fa00660
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4904875
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>