hiberman: Fix hang in dmesg invocation
During suspend we run dmesg to extract the number of dropped pages
with all zeroes from the kernel log. So far we invoked dmesg with
'--since 1 minute ago' to make sure the latest value is extracted,
and not one from a prior hibernation. It turns out that dmesg
often hangs when the logs are limited in that way on an (almost)
hibernated system, more specifically when dmesg tries to open
/etc/localtime. This is probably due to a filesystem or storage
kernel thread that was frozen for hibernation.

Read the entire log instead of limiting the entries with dmesg
parameters and return the last found value.

BUG=b:302568139
TEST=hibernate and resume
  chrome://histograms/droppedpages has a reasonable value

Change-Id: I19f5e650ab692ec2789ae35a460b03be832ff713
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4916634
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>