chromeos-config: Remove hardware-properties codegen check
A long time ago, in a codebase not so far away, cros_config_schema
used to generate C code to compile-in to the EC for hardware
properties.  This no longer exists, but this vestigial check that made
sure the types could be translated to C code still exists.  Drop it.

BUG=none
TEST=CQ

Change-Id: Ib14fd95acaf325f2d38e0a75430ecb992448bed7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4917033
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>