missive: Fix extraneous disk space discard
It happened in rare cases that the file failed to delete, but space was
discarded nevertheless, and then the file (that stayed alive) was
attempted to discard again. The case is rare, so the error did not
happen much. At least that's the only thing I can think of.

Also corrected an error that did not pass desired space to reserve - so
degradation effectively reduced not enough (so we still did see UMA
tracking degradation).

And did some unused stuff removal, minor clean-up.

BUG=b:299372134
TEST=cros_sdk cros_workon_make --board=${BOARD} missive --test

Change-Id: Ic2bfda2928315939ce35b97c038dda8f1d2d9e8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4916635
Reviewed-by: Josh Hilke <jrhilke@google.com>
Commit-Queue: Josh Hilke <jrhilke@google.com>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Auto-Submit: Leonid Baraz <lbaraz@chromium.org>