shill: cellular: Add DEFAULT+DUN metric
Report the correct value in connection_apn_types when using multi PDNs
for tethering. In this case, the metric should report 'Default,Dun' to
indicate that the DUN connection is created while the existing Default
connection remains connected.

BUG=b:291786356
TEST=ff_debug cellular+dbus+device+manager+metrics+modem
ff_debug --level -3
Check the metrics log, when connecting with Dun as Default, Dun+Default,
and Default only.

Change-Id: I32c28166f8de349523048c4f7e2127a760210ff7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4927825
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>