power: Fix cpufreq_config for Adaptive Charging
cpufreq_config is only used on CML, but we need to ensure that we don't
set the battery cpufreq governor when Adaptive Charging is delaying
charge. This does that by not setting the battery governor or other
settings if a full power charger is connected. This will still ensure
that low power settings are used for low power USB chargers, if the
battery starts to discharge.

BUG=b:305276606
TEST=check cpufreq scaling_governor on hatch with and without AC

Change-Id: Ifec6d060460caaca011d3e9317ac3cbcd28419c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4939616
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Yanye Li <yanyeli@google.com>