flex_hwis: use server information based on the flag
The API key and URL on the server side are stored
in the hwis private package. The client needs to add
logic to determine whether the flex internal USE flag
is enabled. If the flag is enabled, the server-related
information can be imported normally from private package.

BUG=b:297887750
TEST=cros_workon_make --board=reven chromeos-base/flex_hwis --test

Change-Id: I476b0c47eebf4b1401f35b614d10a10940d38790
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4923172
Tested-by: Ting-Hao Lin <tinghaolin@google.com>
Commit-Queue: Ting-Hao Lin <tinghaolin@google.com>
Reviewed-by: Cindy Lin <xcl@google.com>
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>