minios: save recovery installer logs
In many autotests, NBR fails at the `RepartitionDisk` step. We don't
have much insight on why this failure occurs since the output of the
`chromeos-install` process is not saved. Save the output to a log file
so that we can inspect it later for any failures/errors.

BUG=b:281583934
TEST=Ran on dut and inspected files.

Change-Id: Ib6f166f59d90dec906a522cedd6f7cedce63d7d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4929048
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Saketh Pothireddy <spothire@google.com>
Tested-by: Saketh Pothireddy <spothire@google.com>