camera: Move camera diagnostics client to hal_adapter
Currently, CameraDiagnosticsClient lives at camera/common
as a singleton instance. It poses architectural challenges
if we want to actively communicate with CameraHalAdapter, or
if we want the diagnostics service to enable/disable frame
interception.

Move the client to hal_adapter and refactor it to make it
easier to access CameraHalAdapter.

BUG=b:253545976
TEST=emerge and deploy cros-camera cros-camera-libs
cros-camera-diagnostics

Cq-Depend: chromium:4891292
Change-Id: Idc58e09d5316080e2b7add697137fdfd138bd878
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4891231
Reviewed-by: Wei Lee <wtlee@chromium.org>
Commit-Queue: Imran Ziad <imranziad@chromium.org>
Tested-by: Imran Ziad <imranziad@chromium.org>