shill: RoutingTable: FlushRoutesWithTag also flush tag-associated table
This is to workaround the behavior that when delete->add a same route,
RoutingTable RTNL listener will capture the delete->add sequence after
RoutingTable add the route cache back, therefore will perform another
round of delete->add, and this will add the route back as tag -1. The
current code will fail to flush this route cache when asked to
FlushRoutesWithTag().

The ideal way to address this is to remove the entire RTNL monitor
logic, which should be no longer needed with accept_ra_rt_table as all
routes should be either direct added by shill or added by kernel and not
cared by shill. This patch is uploaded as a minimum viable solution to
address b/303315643.

BUG=b:303315643
TEST=unit
TEST=manually verify changing address without changing gateway
     no longer triggers b/303315643

Change-Id: Iadf8e0b14d6ea0dbf771c1413086646462ee6722
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4931350
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Taoyu Li <taoyl@chromium.org>