shill: cellular: ignore modem state updates to "connecting"
There is no need to monitor the modem getting into "connecting" state
to update the shill service state to "associating", because this
service state update already happened just before we launched the
connection attempt in shill.

The purpose of this fix was really to avoid a bogus service state
transition to "Associating" whenever the modem was being connected
with direct Simple.Connect() operations without involving shill.
Instead of making the logic more complex, e.g. to check whether a
connection attempt triggered by shill is ongoing, we just remove the
service state update unconditionally in this case, as it really is not
needed for anything.

BUG=b:299669599
TEST=Run in DUT with Verizon+L850:
  tast --verbose run localhost:2201 cellular.IsConnected

Change-Id: Ib7417769e575e6df0db90522ea4aca1363027536
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4932024
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>