hiberman: Fix out of bounds access in hiberlog
hiberlog::replay_line() currently assumes that all log lines start
with '<priority+facility>', in particular it expects that a line
contains'>'. If that isn't the case a panic occurs due to an out
of bounds access in a string. Verify that the string as the
minimum required length before trying to extract a substring.

BUG=b:304598265
TEST=hibernate and resume
  => log replay still works

Change-Id: I95be32318b8a59377af97668e04c56592ca1b059
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4932354
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>