flex_hwis: Stub in flex_disk_metrics source set
This adds an empty pair of flex_disk_metrics source/header files. This
is where most of the implementation will go, and is in a source set so
that it can be shared between the main binary and the test binary.

Update BUILD.gn to add this source set, and include the new header file
in each of the source files.

BUG=b:303112627
TEST=cros_workon_make -breven --test chromeos-base/flex_hwis

Change-Id: I9d7e33a97108e2c19a16b85f028892a09192c20f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4930665
Reviewed-by: Ting-Hao Lin <tinghaolin@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>