flex_hwis: Implement the main binary for flex_disk_metrics
This code delegates to the previously-implemented `GetPartitionSizeMap`
and `SendDiskLayoutMetrics` for most of the work.

The new `GetRootDiskDeviceName` is implemented via the rootdev library,
which is added as a dependency (see also the linked chromiumos-overlay
CL).

The main function now includes the actual list of partitions for which
to send metrics.

Also enabled logging to stderr/syslog.

Corresponding browser CL to define the histogram: crrev.com/c/4936033

BUG=b:303112627
TEST=emerge-reven chromeos-base/flex_hwis
TEST=cros deploy vm chromeos-base/flex_hwis
TEST=Run flex_disk_metrics on VM, verify it exits 0
TEST=Check that the new metrics appear under chrome://histograms

Cq-Depend: chromium:4916525
Change-Id: Id936035d387e1e8ce50fafed1b8418b538231703
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4932845
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Ting-Hao Lin <tinghaolin@google.com>