shill: tethering: Use granular reporting of Entitlement status
Use the new entitlement status values to report when tethering is not
allowed for the user, or when it's not allowed on the carrier.

BUG=b:303901122
TEST=modify the modb to disable tethering on the callbox MNO. Verify
that the correct value is returned when calling CheckReadinessStatus.
Configure the callbox to block the user from tethering. Verify the
entitlemnt check result.

Cq-Depend: chromium:4935406
Change-Id: Iafb6de34adc4881f31a4b0aee820e06ff81e9542
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4937050
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>